Revert "v0.27.0 - new command line arguments"
