+++ exotic/version.py
@@ -1,1 +1,1 @@
-__version__ = "0.27.0"
+__version__ = "0.26.0"
