+++ inits.json
@@ -9,14 +9,12 @@
             "Image Calibrations Directory Guide": "Enter in the path to image calibrations or enter in null for none.",
             "Planetary Parameters Guide": "For planetary parameters that are not filled in, enter in null.",
             "Comparison Star(s) Guide": "Up to 10 comparison stars can be added following the format given below.",
-            "Obs. Latitude Guide": "Indicate the sign (+ North, - South) before the degrees. Needs to be in decimal or HH:MM:SS format.",
-            "Obs. Longitude Guide": "Indicate the sign (+ East, - West) before the degrees. Needs to be in decimal or HH:MM:SS format.",
+            "Obs. Latitude Guide": "Indicate the sign (+ North, - South) before the degrees.",
+            "Obs. Longitude Guide": "Indicate the sign (+ East, - West) before the degrees.",
             "Plate Solution": "For your image to be given a plate solution, type y.",
             "Plate Solution Disclaimer": "One of your imaging files will be publicly viewable on nova.astrometry.net.",
             "Standard Filter": "To use EXOTIC standard filters, type only the filter name.",
             "Custom Filter": "To use a custom filter, enter in the FWHM in optional_info.",
-            "Target Star RA": "Must be in HH:MM:SS sexagesimal format.",
-            "Target Star DEC": "Must be in +/-DD:MM:SS sexagesimal format with correct sign at the beginning (+ or -).",
             "Formatting of null": "Due to the file being a .json, null is case sensitive and must be spelled as shown.",
             "Decimal Format": "Leading zero must be included when appropriate (Ex: 0.32, .32 or 00.32 causes errors.)."
     },
@@ -31,8 +29,8 @@
             "Secondary Observer Codes (N/A if none)": "N/A",
 
             "Observation date": "December 17, 2017",
-            "Obs. Latitude": "+32.41638889",
-            "Obs. Longitude": "-110.73444444",
+            "Obs. Latitude": "+31.68",
+            "Obs. Longitude": "-110.88",
             "Obs. Elevation (meters)": 2616,
             "Camera Type (CCD or DSLR)": "CCD",
             "Pixel Binning": "1x1",
