+++ requirements.txt
@@ -15,11 +15,10 @@ pandas~=1.1
 panel~=0.9
 photutils>=0.7.2
 python_dateutil~=2.8.1
-pyvo~=1.1
 requests~=2.23
 scikit-image~=0.17.2
 scipy~=1.4
 semantic_version~=2.8.5
 setuptools~=50.3.0
 tenacity~=5.1
-wheel~=0.35.1
+pyvo~=1.1
